Fix: Input accessory view theme not updating
The input accessory view was not updating its color theme when the main view's theme changed. This was because the UITextView was not reloading its input views when the inputAccessoryView was reassigned.

This change adds a call to reloadInputViews() after reassigning the inputAccessoryView to force it to reload and reflect the new theme.